+++ src/components/pages/login/LoginPage.js
@@ -1,18 +1,7 @@
-import { useState } from "react";
+import LoginForm from "./LoginForm";
 
 const LoginPage = () => {
-    // state[data, données]
-    const [ inputValue, setInputValue ] = useState("");
-
-    //comportements 
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        alert(`Bonjour ${inputValue}`)
-        setInputValue("")
-    }
-    const handleChange = (event) => {
-        setInputValue(event.target.value)
-    }
+    
 
     //affichage
     // creation du formulaire
@@ -20,12 +9,7 @@ const LoginPage = () => {
     //soumission du formulaire
     return ( 
         <div>
-            <h1>Bienvenue chez nous !</h1><br />
-            <h2>Connectez vous</h2>
-            <form action="submit" onSubmit={handleSubmit}>
-                <input value={inputValue} type="text" placeholder="Entre votre prénom..." required onChange={handleChange} />
-                <button>Accédez à votre page</button>
-            </form>
+            <LoginForm/>
         </div>
      );
 }
